backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.129.0
fire==0.7.1
fonttools==4.61.1
h11==0.16.0
idna==3.11
lxml==6.0.2
numpy==2.4.2
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pdf2docx==0.5.9
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
pypdf==6.7.0
PyPDF2==3.0.1
pypdfium2==5.4.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.22
reportlab==4.4.10
six==1.17.0
starlette==0.52.1
tabula-py==2.10.0
termcolor==3.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
